requirements_local.txt
# Requirements for running the local web UI
gradio==4.44.0
Pillow==10.3.0
gtts==2.5.0
pydub==0.25.1
numpy==2.0.1
opencv-python-headless
librosa
tqdm
imageio
imageio-ffmpeg
scikit-image
